+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v12.1.0
 
 - Using new screenshots grid lazy loading for moderation
+- Prevent users from editing approved screenshots
 - bugfixes
 
 # v12.0.0 - Search screenshots in user space, Shortcuts, dependencies upgrade